+++ src/instruction/swo.cpp
@@ -4,8 +4,7 @@
 Simulator::PreState Simulator::swo(Instruction inst)
 {
     auto op = decodeR(inst);
-    auto addr
-        = static_cast<int32_t>(m_reg.at(op.rt) + m_reg.at(op.rd)) / 4;
+    auto addr = (m_reg.at(op.rt) + m_reg.at(op.rd)) / 4;
     checkMemoryIndex(addr);
 
     auto pre_state = makeMemPreState(addr);
